+++ src/Network/MineCraft/Pi/Client/Internal.hs
@@ -26,12 +26,16 @@ module Network.MineCraft.Pi.Client.Internal
 
 import qualified Control.Exception as CE
 
+import Control.Concurrent (threadDelay)
 import Control.Exception (bracket)
 import Control.Monad (when)
 import Control.Monad.IO.Class
 import Control.Monad.Trans.Reader
 
 import Data.List (intercalate)
+
+import Foreign
+import Foreign.C.String
 
 import Network.BSD
 import Network.Socket
@@ -48,10 +52,18 @@ import System.IO
 type MCPI = ReaderT ConnInfo IO
 
 -- | Connection information.
+--
+-- /TODO:/ Should the buffer used by @flushChannel@ be stored here, to
+--         avoid repeated allocation/de-allocation?  It is unlikely (I
+--         speculate) to be a major optimisation in time or space.
+--
 data ConnInfo = ConnInfo {
       _ciHandle :: Handle  -- ^ Connection to the MineCraft program
     , _ciDebug :: Bool     -- ^ Should messages to and from MineCraft
                            --   be printed to @stderr@.
+    , _ciDelay :: Int      -- ^ Delay, in microseconds, to wait after
+                           --   making a call with `command` before
+                           --   returning. *Experimental*
     }
 
 -- | Commands do not return anything, queries do.
@@ -69,9 +81,10 @@ mcPort = "4711"
 --   /TODO:/ Change the retun value to @Maybe ConnInfo@ and
 --           make this public.
 openMCPI ::
-    Bool  -- ^ Set to @True@ to get debugging messages printed to @stderr@.
-    -> IO ConnInfo
-openMCPI flag = do
+  Int      -- ^ Delay, in microseconds, to wait after making a `command`.
+  -> Bool  -- ^ Set to @True@ to get debugging messages printed to @stderr@.
+  -> IO ConnInfo
+openMCPI delay flag = do
     let ehdl :: CE.IOException -> IO ()
         ehdl _ = 
            hPutStrLn stderr "ERROR: Unable to connect to MineCraft-PI. Is it running?" >>
@@ -81,11 +94,12 @@ openMCPI flag = do
     let a = head as -- note: getAddrInfo never returns an empty list
     sock <- socket (addrFamily a) Stream defaultProtocol
     setSocketOption sock KeepAlive 1
+    setSocketOption sock NoDelay 1
     connect sock (addrAddress a) `CE.catch` ehdl
     h <- socketToHandle sock ReadWriteMode
     hSetBuffering h LineBuffering
 
-    return $ ConnInfo h flag
+    return $ ConnInfo h flag delay
 
 -- | Close the connection.
 --
@@ -93,72 +107,107 @@ openMCPI flag = do
 closeMCPI :: ConnInfo -> IO ()
 closeMCPI = hClose . _ciHandle
 
-logMsg :: ConnInfo -> String -> String -> MCPI ()
+-- | Write a debug message to @stderr@ if the debug flag is set. The
+--   format is \"[type] msg\".
+--
+logMsg ::
+  ConnInfo
+  -> String  -- ^ The type of message (should be 8 characters or less to
+             --   keep the output aligned, but can be larger.
+  -> String  -- ^ The message to display.
+  -> IO ()
 logMsg ConnInfo {..} hdr msg = 
-  when _ciDebug $ liftIO $ hPutStrLn stderr 
+  when _ciDebug $ hPutStrLn stderr 
                 $ "[" ++ hdr ++ replicate (8 - length hdr) ' ' ++
                   "] " ++ msg
 
--- It might be nice to do the argument marshalling here, i.e. have
--- something like @command :: Command -> [Argument] -> MCPI ()@
--- which would be run like @command "player.setTile" [Pos 0 0 0]@,
--- but I do not want to deal with heterogeneous lists at this time.
--- Instead, we force the caller to do the conversion.
+-- | It might be nice to do the argument marshalling here, i.e. have
+--   something like @command :: Command -> [Argument] -> MCPI ()@
+--   which would be run like @command "player.setTile" [Pos 0 0 0]@,
+--   but I do not want to deal with heterogeneous lists at this time.
+--   Instead, we force the caller to do the conversion.
 --
 addArgs :: String -> [Argument] -> String
 addArgs a bs = a ++ "(" ++ intercalate "," bs ++ ")"
 
-{-
+-- | At present the only "error" message I have seen is "Fail\0",
+--   so we use a small buffer size.
+bufSize :: Int
+bufSize = 16
+
+-- | Return from MineCraft that indicates an error.
+connectionError :: String
+connectionError = "Fail"
+
 -- | Remove any output from the handle. A debug message is written
 --   to let the user know this has happened, presumably because of
 --   a previous invalid call. The previous call could be stored in
 --   ConnInfo to make the message more useful, but leave that for now.
 --
 flushChannel :: ConnInfo -> IO ()
-flushChannel ConnInfo {..} = do
+flushChannel ci@ConnInfo {..} = do
+  -- Should the buffer be stored in ConnInfo so we don't have
+  -- to repeatedly allocate it? It's only small (at present),
+  -- so probably not an issue.
+  fbuf <- mallocForeignPtrBytes bufSize 
+  withForeignPtr fbuf $ \bufPtr -> 
+    -- No attempt to reconstruct the message if it requires
+    -- multiple calls. Could be easily added.
+    let loop = do
+          nrec <- hGetBufNonBlocking _ciHandle bufPtr bufSize
+          when (nrec > 0) $ do
+            str <- peekCStringLen (bufPtr, nrec)
+            logMsg ci "FLUSH" str
+            loop
+    in loop
 
-Hmm, do I need the socket as well as the handle?
-
-        while True:
-            readable, _, _ = select.select([self.socket], [], [], 0.0)
-            if not readable:
-                break
-            data = self.socket.recv(1500)
-            e =  "Drained Data: <%s>\n"%data.strip()
-            e += "Last Message: <%s>\n"%self.lastSent.strip()
-            sys.stderr.write(e)
-
--}
-
--- | Run a MineCraft command. 
+-- | Run a MineCraft command. There is a pause after making the command,
+--   which is an attempt to allow any invalid commands to return a
+--   message, so that it can be handled by the next call of `command`
+--   or `query`. This approach is *experimental* and may change.
 command :: Command -> [Argument] -> MCPI ()
 command comm args = do
   ci <- ask
   let commstr = addArgs comm args
-  logMsg ci "COMMAND" commstr
-  liftIO $ hPutStrLn (_ciHandle ci) commstr
+  liftIO $ 
+    flushChannel ci
+    >> logMsg ci "COMMAND" commstr
+    >> hPutStrLn (_ciHandle ci) commstr
+    >> threadDelay (_ciDelay ci)
 
--- | Run a MineCraft query, returning the response.
+-- | Run a MineCraft query, returning the response. An
+--   @IOError@ is raised if the response is @Fail@ (this
+--   might be due to a previous command failing, depending
+--   on the time taken by MineCraft to respond).
+--
 query :: Query -> [Argument] -> MCPI String
 query qry args = do
   ci <- ask
-  let qrystr = addArgs qry args
-  logMsg ci "QUERY" qrystr
-  liftIO $ hPutStrLn (_ciHandle ci) qrystr
+  let qryStr = addArgs qry args
+  liftIO $ 
+    flushChannel ci
+    >> logMsg ci "QUERY" qryStr
+    >> hPutStrLn (_ciHandle ci) qryStr
   ans <- liftIO $ hGetLine (_ciHandle ci)
-  logMsg ci "RESPONSE" ans
+  liftIO $ logMsg ci "RESPONSE" ans
+  when (ans == connectionError) $
+    liftIO $ ioError $ userError $ "Query failed: " ++ qryStr
   return ans
+
+-- | Try a 0.1 second delay.
+testDelay :: Int
+testDelay = 100000
 
 -- | Run a Raspberry-PI program. The flag determines whether the
 --   messages sent to, and received from, the server, are
---   printed to @stderr@.
+--   printed to @stderr@ as a diagnostic.
 --
 --   An exception is raised if the server is not running, or
 --   can not be contacted.
 runMCPI' :: Bool -> MCPI a -> IO a
 runMCPI' flag p = 
   bracket
-    (openMCPI flag)
+    (openMCPI testDelay flag)
     closeMCPI
     (runReaderT p)